+++ integration/token/fungible/views/checks.go
@@ -117,7 +117,7 @@ func (m *CheckTTXDBView) Call(context view.Context) (interface{}, error) {
 			if err != nil {
 				errorMessages = append(errorMessages, fmt.Sprintf("failed to get ledger transaction status for [%s]", transactionRecord.TxID))
 			}
-			errorMessages = append(errorMessages, fmt.Sprintf("transaction record [%s] is unknwon for vault but not for the ledger [%d]", transactionRecord.TxID, lVC))
+			errorMessages = append(errorMessages, fmt.Sprintf("transaction record [%s] is unknown for vault but not for the ledger [%d]", transactionRecord.TxID, lVC))
 		case vc == network.Busy && lVC != network.Unknown:
 			if err != nil {
 				errorMessages = append(errorMessages, fmt.Sprintf("failed to get ledger transaction status for [%s]", transactionRecord.TxID))
